Implemented quicksort.
This commit implements quicksort, with both an inner helper function
and a separate partition function.

This quicksort implementation is fairly simple, and always uses
the first element in a range as the pivot. There are smarter pivot-
selection schemes to choose from, but for the moment this will do
just fine.