Modified quick_select to work on int over char.
The original version was a direct adaptation from the quick_sort
code, which operated on buffers of chars. Quick select is really
more sensible with integers, so that's what it works on now.

This required making integer versions of some of the underlying
functions. Theoretically, since chars are integers, these could
be unified. For now they're left distinct.